Initial state for Lesson 4
And for initial, I mean already working as intended. Kind of.
I still need to remember precisely what was the effect to be
achieved. Also, I want to refactor and optmize some things.

Please notice that the ticker thread was done from Java rather
than C++, so the Framework has greater control over the
rendering code.